Optix Tech Test
Used the template provided and implemented the requested features.

I had trouble with two of the requested features:

1. Sorting the review column. There is no API to get the reviews, so I
didn't add a review column and therefore it is not sortable.

2. Getting the review box to display in a modal on mobile devices. I
wasn't able to get the modal to show up at all, let alone on mobile. If
I spent some more time investigating I would be able to solve this one.